requirements_2.txt
azure-cognitiveservices-speech
azure-identity
flask
flask-socketio
openai
pytz
requests
torch
numpy
torchaudio
gunicorn
eventlet